+++ src/Exception/InvalidDynamicValuesClosureException.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Sofascore\PurgatoryBundle\Exception;
 
-final class InvalidIfClosureException extends InvalidArgumentException
+final class InvalidDynamicValuesClosureException extends InvalidArgumentException
 {
-    private const MESSAGE = 'Invalid "if" closure provided for route "%s": "%s"';
+    private const MESSAGE = 'Invalid "DynamicValues" closure provided for route "%s": "%s"';
 
     public function __construct(
         public readonly string $routeName,